Fix extension on EDB's fork of postgres.
An unqualified usage of "rownum" leads to the Oracle-specific behavior of a
plain LIMIT X, even if there's a attribute named rownum in the visible
relations.

Simply qualifying the rownum field is enough, but I'm also renaming it to "rn"
for extra safety and make sure such usage won't be reintroduced later, as I
don't have a way to test the EDB specific behavior.

Thanks to github user manishnew09 for the report, and both manishnew09 and
Thomas Reiss for helping identifying the root problem.