Fix role / db setting snapshot when all previous settings have been removed
In such case the datasource is empty and there's therefore no assiciated
timestamp to use when recording that the setting was removed.  In that case
simply use "now()".

This was supposed to be covered by the regression tests, but pg_regress creates
the contrib_regression database with specific settings to ensure reproducible
tests.

Thanks to github user sajiljosephs for the report.